Update install scripts
installation of clang-format requires 'optional' option
ninja installed at top level
fix esp32 install
Fix shellcheck globbing warnings
Use arrays consistently
Make non-apple CLANG available for RBPF library
Don't change permissions on ``opt`` as the actual directory could be different (as per environment variables).

+++ Tools/install.sh
@@ -5,7 +5,7 @@
 #    . /opt/sming/Tools/install.sh
 #
 
-[ "$0" = "$BASH_SOURCE" ]; sourced=$?
+[ "$0" = "${BASH_SOURCE[0]}" ]; sourced=$?
 
 inst_host=0
 inst_doc=0
@@ -14,7 +14,16 @@ inst_esp32=0
 inst_rp2040=0
 err=0
 
-FONT_PACKAGES="fonts-ubuntu fonts-noto-mono xfonts-base fonts-urw-base35 fonts-droid-fallback"
+FONT_PACKAGES=(\
+    fonts-ubuntu \
+    fonts-noto-mono \
+    xfonts-base \
+    fonts-urw-base35 \
+    fonts-droid-fallback\
+    )
+
+EXTRA_PACKAGES=()
+OPTIONAL_PACKAGES=()
 
 for opt in "$@"; do
     opt=$(echo "$opt" | tr '[:upper:]' '[:lower:]')
@@ -31,7 +40,11 @@ for opt in "$@"; do
             ;;
 
         fonts)
-            EXTRA_PACKAGES+=" $FONT_PACKAGES"
+            EXTRA_PACKAGES+=("${FONT_PACKAGES[@]}")
+            ;;
+
+        optional)
+            OPTIONAL_PACKAGES+=(clang-format-8)
             ;;
 
         *)
@@ -50,6 +63,7 @@ if [[ $err -eq 1 ]] || [ $# -eq 0 ]; then
     echo '   all       Install all architectures'
     echo '   doc       Tools required to build documentation'
     echo '   fonts     Install fonts used by Graphics library (normally included with Ubuntu)'
+    echo '   optional  Install optional development tools'
     echo
     if [ $sourced = 1 ]; then
         return 1
@@ -59,25 +73,22 @@ if [[ $err -eq 1 ]] || [ $# -eq 0 ]; then
 fi
 
 # Sming repository for binary archives
-SMINGTOOLS=https://github.com/SmingHub/SmingTools/releases/download/1.0
+export SMINGTOOLS=https://github.com/SmingHub/SmingTools/releases/download/1.0
 
 # Set default environment variables
-if [ -z "$APPVEYOR" ]; then
-    source $(dirname "$BASH_SOURCE")/export.sh
+source "$(dirname "${BASH_SOURCE[0]}")/export.sh"
 
-    # Ensure default path is writeable
-    sudo mkdir -p /opt
-    sudo chown $USER:$USER /opt
-fi
-
-WGET="wget --no-verbose"
+export WGET="wget --no-verbose"
 
 # Installers put downloaded archives here
 DOWNLOADS="downloads"
 mkdir -p $DOWNLOADS
 
 # Identify package installer for distribution
-if [ -n "$(command -v apt)" ]; then
+if [[ "$(uname)" = "Darwin" ]]; then
+    DIST=darwin
+    PKG_INSTALL="brew install"
+elif [ -n "$(command -v apt)" ]; then
     DIST=debian
     PKG_INSTALL="sudo apt-get install -y"
 elif [ -n "$(command -v dnf)" ]; then
@@ -94,30 +105,45 @@ fi
 
 # Common install
 
-if [ -n "$APPVEYOR" ] || [ -n "$GITHUB_ACTION" ]; then
+if [ -n "$GITHUB_ACTIONS" ]; then
 
-    # Provide repo. for clang-format-8 on Ubuntu 22.04
-    sudo apt-add-repository -y 'deb http://mirrors.kernel.org/ubuntu focal main universe'
-    sudo apt-get -y update
-    $PKG_INSTALL \
-        clang-format-8 \
-        g++-multilib \
-        python3-setuptools \
-        ninja-build \
-        linux-modules-extra-$(uname -r) \
-        exfatprogs \
-        $EXTRA_PACKAGES
+    case $DIST in
+        debian)
+            if [ ${#OPTIONAL_PACKAGES[@]} ]; then
+                # Provide repo. for clang-format-8 on Ubuntu 22.04
+                sudo apt-add-repository -y 'deb http://mirrors.kernel.org/ubuntu focal main universe'
+            fi
+            sudo apt-get -y update
+            $PKG_INSTALL \
+                g++-multilib \
+                python3-setuptools \
+                ninja-build \
+                "linux-modules-extra-$(uname -r)" \
+                exfatprogs \
+                "${EXTRA_PACKAGES[@]}" \
+                "${OPTIONAL_PACKAGES[@]}"
+            ;;
+
+        darwin)
+            $PKG_INSTALL \
+                binutils \
+                coreutils \
+                gnu-sed \
+                ninja
+            ;;
+
+        esac
 
 else
 
-    MACHINE_PACKAGES=""
+    MACHINE_PACKAGES=()
     case $DIST in
         debian)
             case $(uname -m) in
                 arm | aarch64)
                     ;;
                 *)
-                    MACHINE_PACKAGES="g++-multilib"
+                    MACHINE_PACKAGES+=(g++-multilib)
                     ;;
             esac
             sudo apt-get -y update || echo "Update failed... Try to install anyway..."
@@ -133,16 +159,21 @@ else
             	python3-pip \
             	python3-setuptools \
                 wget \
-                $MACHINE_PACKAGES \
-                $EXTRA_PACKAGES
+                "${MACHINE_PACKAGES[@]}" \
+                "${EXTRA_PACKAGES[@]}"
 
-            $PKG_INSTALL clang-format-8 || printf "\nWARNING: Failed to install optional clang-format-8.\n\n"
+            if [ -n "$OPTIONAL_PACKAGES" ]; then
+                $PKG_INSTALL "${OPTIONAL_PACKAGES[@]}" || printf "\nWARNING: Failed to install optional %s.\n\n" "$OPTIONAL_PACKAGES"
+            fi
             ;;
 
         fedora)
             case $(uname -m) in
                 x86_64)
-                    MACHINE_PACKAGES="glibc-devel.i686 libstdc++.i686"
+                    MACHINE_PACKAGES=(\
+                        glibc-devel.i686 \
+                        libstdc++.i686 \
+                        )
                     ;;
             esac
             $PKG_INSTALL \
@@ -159,12 +190,22 @@ else
                 sed \
                 unzip \
                 wget \
-                $MACHINE_PACKAGES
+                "${MACHINE_PACKAGES[@]}"
+            ;;
+
+        darwin)
+            $PKG_INSTALL \
+                binutils \
+                coreutils \
+                gnu-sed \
+                ninja
             ;;
 
     esac
 
-    sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 100
+    if [ "$DIST" != "darwin" ]; then
+        sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 100
+    fi
 
 fi
 
@@ -206,7 +247,7 @@ if [ $inst_rp2040 -eq 1 ]; then
 fi
 
 if [ -z "$KEEP_DOWNLOADS" ]; then
-    rm -rf "$DOWNLOADS/*"
+    rm -rf "${DOWNLOADS:?}/"*
 fi
